-Created some more platform types -Added a bouncy platform, that gives you twice your jumpheight -Added a unbouncy platform, which stops your movespeed and vertical momentum. -Shifted a few things around on the test level
The platforms are still a little iffy, but I think thats more to do with aligning the hitboxes correctly, so I'm not too worried.
I've also added some really basic assets for them, and they're terrible and need to be replaced
I've also for an asset for some moving platforms, though I haven't started work on those yet.